requirements.txt
appdirs==1.4.3
asgiref==3.3.4
attrs==21.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.2.4
django-autoslug==1.9.8
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-heroku==0.3.1
django-phonenumber-field==5.2.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.0
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
drf-spectacular==0.17.2
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
lockfile==0.12.2
Markdown==3.3.4
MarkupSafe==2.0.1
msgpack==0.6.2
oauthlib==3.1.1
packaging==20.3
pep517==0.8.2
phonenumbers==8.12.25
Pillow==8.3.0
progress==1.5
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.6
pyrsistent==0.18.0
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
spectacular==0.4.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
